importruns: extract run path traversal logic to separate concerns
Move the run path traversal logic out of the importruns management
command into a separate module to reduce complexity and separate
concerns.
